urlDict: move to dict folder
